% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter-registry.R
\name{adapter_registry}
\alias{adapter_registry}
\title{List all registered adapters}
\usage{
adapter_registry()
}
\value{
a named list of adapters, keyed by adapter \code{id}.
}
\description{
List all registered adapters
}
