% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{.recycle_units}
\alias{.recycle_units}
\title{Recycle a units vector to a target length, erroring on a bad length}
\usage{
.recycle_units(units_x, n, arg_name)
}
\description{
Recycle a units vector to a target length, erroring on a bad length
}
\keyword{internal}
