% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{.resolve_lab_datetime_formats}
\alias{.resolve_lab_datetime_formats}
\title{Resolve a \code{formats} argument to a vector of \code{strptime()} format strings}
\usage{
.resolve_lab_datetime_formats(formats)
}
\description{
Resolve a \code{formats} argument to a vector of \code{strptime()} format strings
}
\keyword{internal}
